DAP-127: Adding rlwrap support
This also moves to always running the setup script in a docker
container, which simplifies the script and ensures we can satisfy any
dependencies we need.

A result of this is that the build of the setup-env image has moved to
an internal repo with the image published to dockerhub.